LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-position/sticky/position-sticky-scrollIntoView-expected.txt

Title 1

Title 2

Title 3


FAIL scrolling a sticky element into view should use its unshifted position assert_approx_equals: expected 1120 +/- 1 but got 560
